requirements.txt
# local package
-e .

# external requirements
flake8
cookiecutter
dvc
pandas
numpy
matplotlib
seaborn
scikit-learn
joblib
flask
pyyaml
mlflow==1.27.0